Factor PlatformIO buildgen out of writer.py
